feat: 0.5점 별점 · 감상/위시리스트 수정 · 읽고 싶은 책 페이지 추가
- 감상 별점을 0.5 단위로 입력/표시 (rating Int→Float)
- 감상에 수정 버튼 추가 (폼 자동 채움)
- 읽고 싶은 책 리스트 페이지(/wishlist) 신설 — 기획/시스템/사람/기타 분류, 추가·수정·삭제
- Supabase directUrl 설정 추가
- 확장 프로그램 하이드레이션 경고 억제(suppressHydrationWarning)

+++ app/actions/reviews.ts
@@ -7,7 +7,7 @@ export async function upsertReview(formData: FormData) {
     const bookId = parseInt(formData.get("bookId") as string);
     const memberId = parseInt(formData.get("memberId") as string);
     const content = formData.get("content") as string;
-    const rating = parseInt(formData.get("rating") as string);
+    const rating = parseFloat(formData.get("rating") as string);
 
     await prisma.review.upsert({
         where: { bookId_memberId: { bookId, memberId } },
